completed of exception handling

+++ src/main/java/com/example/rest/demo/service/useService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.rest.demo.repository.UserRepository;
 
+import ExceptionHandling.ServiceException;
 import jakarta.persistence.Id;
 
 import com.example.rest.demo.entities.User;
@@ -21,27 +22,43 @@ public class useService {
 		return rep.findAll();		
 		
 	}
-	public User saveall(User user) {
+	public User saveall(User user) throws ServiceException  {
+		Optional <User> existName= rep.findByFirstName(user.getFirstName());
+		if(existName.isPresent()) {
+			throw new ServiceException("there is same name on database");
+		}
 		return rep.save(user);
 		
 	}
-	public Optional<User> userid(Long id){
+	public Optional<User> userid(Long id) throws ServiceException {
 		Optional<User>user = rep.findById(id);
+		if(!user.isPresent()) {
+			throw new ServiceException("there is no id on it");
+		}
 		return user;
 	}
     
-	public User updateId (Long id , User user) {
+	public User updateId (Long id , User user)   throws ServiceException{
+		
 		user.setSno(id);
+		Optional<User>userUpdate = rep.findById(id);
+		if(!userUpdate.isPresent()) {
+			throw new ServiceException("we cant update cause there no id");
+		}
 		return rep.save(user);		
 	}
 	
-	public void deleteId (Long id) {
+	public void deleteId (Long id) throws ServiceException {
+		Optional<User>userDelete = rep.findById(id);
+		if(!userDelete.isPresent()) {
+			throw new ServiceException("we cant delete cause there no id");
+		}
 		if(rep.findById(id).isPresent()) {
 			rep.deleteById(id);
 		}
 	}
 	
-	public List<User> findusername(String firstName) {
+	public Optional<User> findusername(String firstName) {
 		 return rep.findByFirstName(firstName);
 		
 	}
